add map loading (scenery + splines) integrate physx into project ! fixme ! include debug symbols in the weird build configurations nvidia forces onto us

+++ source/scene/transformcomponents.hpp
@@ -19,7 +19,7 @@ namespace simp {
   };
 
   struct TransformWorldComponent {
-    glm::vec2 Tile;
+    glm::ivec2 Tile;
   };
 
   struct TransformParentComponent {
